new test not working, troubleshooting
