+++ api/tests/test_create_order.py
@@ -2,27 +2,12 @@ from fastapi.testclient import TestClient
 from main import app
 from queries.orders import OrderQueries
 
+# This will be challenging because of Mongo automatically adding:
+# id, date, and customer_username to the mock order
+# Current Status: test failing because 401 unauthorized. This likely has to do with having to make a mock login event
+
 
 client = TestClient(app)
-
-
-class EmptyOrderQueries:
-    def list_orders(self):
-        return []
-
-
-def test_list_orders():
-    # Arrange
-    app.dependency_overrides[OrderQueries] = EmptyOrderQueries
-
-    response = client.get("/orders")
-
-    # Act
-    app.dependency_overrides = {}
-
-    # Assert
-    assert response.status_code == 200
-    assert response.json() == []
 
 
 class CreateOrderQueries:
@@ -34,15 +19,12 @@ class CreateOrderQueries:
 
 def test_create_order():
     app.dependency_overrides[OrderQueries] = CreateOrderQueries
-    # this will be challenging because of Mongo automatically adding:
-    # id, date, and customer_username
-    # to the mock order
 
     json = {
         "date": "2024-01-15, 22:14",
         "reviewed": False,
         "order_status": "Order received",
-        "customer_username": "Channel Islands",
+        "customer_username": "KellySlater",
         "surfboard_shaper": "Rusty",
         "surfboard_model": "Twin fin",
         "surfboard_length": 6,
@@ -61,7 +43,7 @@ def test_create_order():
         "date": "2024-01-15, 22:14",
         "reviewed": False,
         "order_status": "Order received",
-        "customer_username": "Channel Islands",
+        "customer_username": "KellySlater",
         "surfboard_shaper": "Rusty",
         "surfboard_model": "Twin fin",
         "surfboard_length": 6,
